+++ src/Components/Dashboard/VMGroup/VM_group.js
@@ -1,23 +1,24 @@
 import React from 'react';
+import { makeStyles, Button } from '@material-ui/core';
 import { useSelector, useDispatch } from 'react-redux';
 import { showClouds } from '../../../Actions/CloudActions';
-import { showVMGroup, getTemplates } from '../../../Actions/VM_groupActions'
+import { showVMGroup, getTemplates } from '../../../Actions/VM_groupActions';
 import { TableOfVMGroup } from './TableOfVMGroups';
-import { makeStyles, Button } from '@material-ui/core';
 import { CreateVMGroup } from './CreateVMGroup';
+import { serverURL } from '../Dashboard';
 
 const useStyles = makeStyles(theme => ({
 	createButtonMargin: {
 		marginBottom: '20px'
 	},
+}));
 
-}))
 export function VM_group() {
 	const classes = useStyles();
 	const dispatch = useDispatch();
 	const templates = useSelector(state => state.templates);
+	const clouds = useSelector(state => state.clouds);
 	const [createVmGroupWindowOpen, setCreateVmGroupWindowOpen] = React.useState(false);
-	const serverURL = 'http://192.168.84.189:8080';
 
 	const refreshCloudData = () => {
 		fetch(`${serverURL}/api/cloud_providers/list`)
@@ -33,10 +34,7 @@ export function VM_group() {
 		fetch(`${serverURL}/api/proxmox/vm/group/list`)
 		.then(response => response.json())
 		.then(data => data.vm_group_list)
-		.then(data => {
-			dispatch(showVMGroup(data));
-			return data;
-		})
+		.then(data => dispatch(showVMGroup(data)))
 		refreshTempaltes();
 	};
 	const refreshTempaltes = () => {
@@ -44,7 +42,7 @@ export function VM_group() {
 		.then(response => response)
 		.then(data => data.json())
 		.then(data => dispatch(getTemplates(data.template_list)))
-	}
+	};
 	const handleCreateVmGroupWindowOpen = () => {
 		refreshCloudData();
 		setCreateVmGroupWindowOpen(true);
@@ -57,12 +55,16 @@ export function VM_group() {
 				createVmGroupWindowOpen={createVmGroupWindowOpen}
 				setCreateVmGroupWindowOpen={setCreateVmGroupWindowOpen}
 				refreshVMGroupData={refreshVMGroupData}
+				dispatch={dispatch}
+				templates={templates}
+				clouds={clouds}
 			/>
 			<TableOfVMGroup
 				refreshVMGroupData={refreshVMGroupData}
 				refreshCloudData={refreshCloudData}
 				templates={templates}
+				clouds={clouds}
 			/>
 		</div>
-	)
-}
+	);
+};
